+++ sdk/network/azure-mgmt-network/azure/mgmt/network/v2018_02_01/_network_management_client.py
@@ -9,20 +9,65 @@
 from copy import deepcopy
 from typing import Any, TYPE_CHECKING
 
-from msrest import Deserializer, Serializer
-
 from azure.core.rest import HttpRequest, HttpResponse
 from azure.mgmt.core import ARMPipelineClient
 
 from . import models
+from .._serialization import Deserializer, Serializer
 from ._configuration import NetworkManagementClientConfiguration
-from .operations import ApplicationGatewaysOperations, ApplicationSecurityGroupsOperations, AvailableEndpointServicesOperations, BgpServiceCommunitiesOperations, ConnectionMonitorsOperations, DdosProtectionPlansOperations, DefaultSecurityRulesOperations, ExpressRouteCircuitAuthorizationsOperations, ExpressRouteCircuitConnectionsOperations, ExpressRouteCircuitPeeringsOperations, ExpressRouteCircuitsOperations, ExpressRouteCrossConnectionPeeringsOperations, ExpressRouteCrossConnectionsOperations, ExpressRouteServiceProvidersOperations, InboundNatRulesOperations, LoadBalancerBackendAddressPoolsOperations, LoadBalancerFrontendIPConfigurationsOperations, LoadBalancerLoadBalancingRulesOperations, LoadBalancerNetworkInterfacesOperations, LoadBalancerProbesOperations, LoadBalancersOperations, LocalNetworkGatewaysOperations, NetworkInterfaceIPConfigurationsOperations, NetworkInterfaceLoadBalancersOperations, NetworkInterfacesOperations, NetworkManagementClientOperationsMixin, NetworkSecurityGroupsOperations, NetworkWatchersOperations, Operations, PacketCapturesOperations, PublicIPAddressesOperations, RouteFilterRulesOperations, RouteFiltersOperations, RouteTablesOperations, RoutesOperations, SecurityRulesOperations, SubnetsOperations, UsagesOperations, VirtualNetworkGatewayConnectionsOperations, VirtualNetworkGatewaysOperations, VirtualNetworkPeeringsOperations, VirtualNetworksOperations
+from .operations import (
+    ApplicationGatewaysOperations,
+    ApplicationSecurityGroupsOperations,
+    AvailableEndpointServicesOperations,
+    BgpServiceCommunitiesOperations,
+    ConnectionMonitorsOperations,
+    DdosProtectionPlansOperations,
+    DefaultSecurityRulesOperations,
+    ExpressRouteCircuitAuthorizationsOperations,
+    ExpressRouteCircuitConnectionsOperations,
+    ExpressRouteCircuitPeeringsOperations,
+    ExpressRouteCircuitsOperations,
+    ExpressRouteCrossConnectionPeeringsOperations,
+    ExpressRouteCrossConnectionsOperations,
+    ExpressRouteServiceProvidersOperations,
+    InboundNatRulesOperations,
+    LoadBalancerBackendAddressPoolsOperations,
+    LoadBalancerFrontendIPConfigurationsOperations,
+    LoadBalancerLoadBalancingRulesOperations,
+    LoadBalancerNetworkInterfacesOperations,
+    LoadBalancerProbesOperations,
+    LoadBalancersOperations,
+    LocalNetworkGatewaysOperations,
+    NetworkInterfaceIPConfigurationsOperations,
+    NetworkInterfaceLoadBalancersOperations,
+    NetworkInterfacesOperations,
+    NetworkManagementClientOperationsMixin,
+    NetworkSecurityGroupsOperations,
+    NetworkWatchersOperations,
+    Operations,
+    PacketCapturesOperations,
+    PublicIPAddressesOperations,
+    RouteFilterRulesOperations,
+    RouteFiltersOperations,
+    RouteTablesOperations,
+    RoutesOperations,
+    SecurityRulesOperations,
+    SubnetsOperations,
+    UsagesOperations,
+    VirtualNetworkGatewayConnectionsOperations,
+    VirtualNetworkGatewaysOperations,
+    VirtualNetworkPeeringsOperations,
+    VirtualNetworksOperations,
+)
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from azure.core.credentials import TokenCredential
 
-class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pylint: disable=too-many-instance-attributes
+
+class NetworkManagementClient(
+    NetworkManagementClientOperationsMixin
+):  # pylint: disable=client-accepts-api-version-keyword,too-many-instance-attributes
     """Network Client.
 
     :ivar application_gateways: ApplicationGatewaysOperations operations
@@ -141,10 +186,10 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
     :ivar local_network_gateways: LocalNetworkGatewaysOperations operations
     :vartype local_network_gateways:
      azure.mgmt.network.v2018_02_01.operations.LocalNetworkGatewaysOperations
-    :param credential: Credential needed for the client to connect to Azure.
+    :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The subscription credentials which uniquely identify the Microsoft
-     Azure subscription. The subscription ID forms part of the URI for every service call.
+     Azure subscription. The subscription ID forms part of the URI for every service call. Required.
     :type subscription_id: str
     :param base_url: Service URL. Default value is "https://management.azure.com".
     :type base_url: str
@@ -159,7 +204,9 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
         base_url: str = "https://management.azure.com",
         **kwargs: Any
     ) -> None:
-        self._config = NetworkManagementClientConfiguration(credential=credential, subscription_id=subscription_id, **kwargs)
+        self._config = NetworkManagementClientConfiguration(
+            credential=credential, subscription_id=subscription_id, **kwargs
+        )
         self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
@@ -199,9 +246,7 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
         self.express_route_cross_connection_peerings = ExpressRouteCrossConnectionPeeringsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.load_balancers = LoadBalancersOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.load_balancers = LoadBalancersOperations(self._client, self._config, self._serialize, self._deserialize)
         self.load_balancer_backend_address_pools = LoadBalancerBackendAddressPoolsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
@@ -232,51 +277,35 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
         self.network_security_groups = NetworkSecurityGroupsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.security_rules = SecurityRulesOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.security_rules = SecurityRulesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.default_security_rules = DefaultSecurityRulesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
         self.network_watchers = NetworkWatchersOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.packet_captures = PacketCapturesOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.packet_captures = PacketCapturesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.connection_monitors = ConnectionMonitorsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.operations = Operations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.operations = Operations(self._client, self._config, self._serialize, self._deserialize)
         self.public_ip_addresses = PublicIPAddressesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.route_filters = RouteFiltersOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.route_filters = RouteFiltersOperations(self._client, self._config, self._serialize, self._deserialize)
         self.route_filter_rules = RouteFilterRulesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.route_tables = RouteTablesOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
-        self.routes = RoutesOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.route_tables = RouteTablesOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.routes = RoutesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.bgp_service_communities = BgpServiceCommunitiesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.usages = UsagesOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.usages = UsagesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.virtual_networks = VirtualNetworksOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.subnets = SubnetsOperations(
-            self._client, self._config, self._serialize, self._deserialize
-        )
+        self.subnets = SubnetsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.virtual_network_peerings = VirtualNetworkPeeringsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
@@ -290,12 +319,7 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
             self._client, self._config, self._serialize, self._deserialize
         )
 
-
-    def _send_request(
-        self,
-        request: HttpRequest,
-        **kwargs: Any
-    ) -> HttpResponse:
+    def _send_request(self, request: HttpRequest, **kwargs: Any) -> HttpResponse:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
@@ -304,7 +328,7 @@ class NetworkManagementClient(NetworkManagementClientOperationsMixin):    # pyli
         >>> response = client._send_request(request)
         <HttpResponse: 200 OK>
 
-        For more information on this code flow, see https://aka.ms/azsdk/python/protocol/quickstart
+        For more information on this code flow, see https://aka.ms/azsdk/dpcodegen/python/send_request
 
         :param request: The network request you want to make. Required.
         :type request: ~azure.core.rest.HttpRequest