+++ tools/azure-sdk-tools/devtools_testutils/proxy_startup.py
@@ -28,9 +28,10 @@ WINDOWS_IMAGE_SOURCE_PREFIX = "azsdkengsys.azurecr.io/engsys/testproxy-win"
 CONTAINER_STARTUP_TIMEOUT = 6000
 PROXY_MANUALLY_STARTED = os.getenv("PROXY_MANUAL_START", False)
 
-PROXY_CHECK_URL = PROXY_URL.rstrip("/") + "/Info/Available"
+PROXY_CHECK_URL = PROXY_URL + "/Info/Available"
 TOOL_ENV_VAR = "PROXY_PID"
 
+discovered_roots = []
 
 def get_image_tag(repo_root: str) -> str:
     """Gets the test proxy Docker image tag from the target_version.txt file in /eng/common/testproxy"""
@@ -59,6 +60,8 @@ def ascend_to_root(start_dir_or_file: str) -> str:
 
         # we need the git check to prevent ascending out of the repo
         if os.path.exists(possible_root):
+            if current_dir not in discovered_roots:
+                discovered_roots.append(current_dir)
             return current_dir
         else:
             current_dir = os.path.dirname(current_dir)
@@ -140,10 +143,15 @@ def start_test_proxy(request) -> None:
                 log = open(os.path.join(root, "_proxy_log_{}.log".format(envname)), "a")
 
                 _LOGGER.info("{} is calculated repo root".format(root))
+
+                os.environ["PROXY_ASSETS_FOLDER"] = os.path.join(root, "l", envname)
+                if not os.path.exists(os.environ["PROXY_ASSETS_FOLDER"]):
+                    os.makedirs(os.environ["PROXY_ASSETS_FOLDER"])
+
                 proc = subprocess.Popen(
                     shlex.split('test-proxy start --storage-location="{}" -- --urls "{}"'.format(root, PROXY_URL)),
                     stdout=log,
-                    stderr=log,
+                    stderr=log
                 )
                 os.environ[TOOL_ENV_VAR] = str(proc.pid)
         else: