+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 12.9.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+- Removed `msrest` dependency.
+- Added `typing-extensions>=4.0.1` as a dependency.
+- Added extra dependency `aio` for installing optional async dependencies. Use `pip install azure-storage-file-datalake[aio]` to install.
+
 ## 12.9.0 (2022-10-11)
 
 ### Features Added